Implements bare-bones of the create and delete order feature. Uses sql_alchemy to store information, and a seperate proccess to handle command line input.
To use SQLAlchemy new classes that were compatible had to be made for item and order. They are currently in PizzaParlour, but should be moved. The CLI defined in Commands.py was made under the assumption that we would do calculations there. We should probably switch to a Click cli since it's not.