task1 is roughly complete. Code is still messy.
Big picture we are using a sqlite database defined in Models.py to store info of a flask server whose behaviour is defined in PizzaParlour.py currently, which we contact using Commands.py, which is just a barebones input program. Specifically got the db and request processing seperate from the app and the test environment. Added some mocking for the html request sent by Commands.